+++ tests/Stash/Test/AbstractPoolTest.php
@@ -101,7 +101,7 @@ class AbstractPoolTest extends \PHPUnit_Framework_TestCase
     {
         $pool = $this->getTestPool();
 
-        $this->assertFalse($pool->exists('base/one'), 'Pool->exists() returns false for item without stored data.');
+        $this->assertFalse($pool->hasItem('base/one'), 'Pool->hasItem() returns false for item without stored data.');
         $item = $pool->getItem('base', 'one');
         $this->assertInstanceOf('Stash\Item', $item, 'getItem returns a Stash\Item object');
 
@@ -117,7 +117,7 @@ class AbstractPoolTest extends \PHPUnit_Framework_TestCase
         $storedData = $item->get();
         $this->assertEquals($this->data, $storedData, 'Pool->save() returns proper data on new Item instance.');
 
-        $this->assertTrue($pool->exists('base/one'), 'Pool->exists() returns true for item with stored data.');
+        $this->assertTrue($pool->hasItem('base/one'), 'Pool->hasItem() returns true for item with stored data.');
 
         $pool->setNamespace('TestNamespace');
         $item = $pool->getItem(array('test', 'item'));
@@ -130,7 +130,7 @@ class AbstractPoolTest extends \PHPUnit_Framework_TestCase
     {
         $pool = $this->getTestPool();
 
-        $this->assertFalse($pool->exists('base/one'), 'Pool->exists() returns false for item without stored data.');
+        $this->assertFalse($pool->hasItem('base/one'), 'Pool->hasItem() returns false for item without stored data.');
         $item = $pool->getItem('base', 'one');
         $this->assertInstanceOf('Stash\Item', $item, 'getItem returns a Stash\Item object');
 
@@ -146,7 +146,7 @@ class AbstractPoolTest extends \PHPUnit_Framework_TestCase
         $storedData = $item->get();
         $this->assertEquals($this->data, $storedData, 'Pool->save() returns proper data on new Item instance.');
 
-        $this->assertTrue($pool->exists('base/one'), 'Pool->exists() returns true for item with stored data.');
+        $this->assertTrue($pool->hasItem('base/one'), 'Pool->hasItem() returns true for item with stored data.');
 
         $pool->setNamespace('TestNamespace');
         $item = $pool->getItem(array('test', 'item'));
@@ -154,14 +154,14 @@ class AbstractPoolTest extends \PHPUnit_Framework_TestCase
         $this->assertAttributeEquals('TestNamespace', 'namespace', $item, 'Pool sets Item namespace.');
     }
 
-    public function testExists()
+    public function testHasItem()
     {
         $pool = $this->getTestPool();
-        $this->assertFalse($pool->exists('base/one'), 'Pool->exists() returns false for item without stored data.');
+        $this->assertFalse($pool->hasItem('base/one'), 'Pool->hasItem() returns false for item without stored data.');
         $item = $pool->getItem('base', 'one');
         $item->set($this->data);
         $pool->save($item);
-        $this->assertTrue($pool->exists('base/one'), 'Pool->exists() returns true for item with stored data.');
+        $this->assertTrue($pool->hasItem('base/one'), 'Pool->hasItem() returns true for item with stored data.');
     }
 
     public function testCommit()